+++ source/BaseSocket.h
@@ -16,8 +16,8 @@ public:
 protected:
     bool open_socket();
     void close_socket(); // for now never used
-    bool send_packet(char buffer[BUFFER_SIZE],uint8 datalength);
-    bool recv_packet(char * buffer,uint8* datalength);
+    bool send_packet(char buffer[BUFFER_SIZE],uint16 datalength);
+    bool recv_packet(char * buffer,uint16* datalength);
 private:
     WSADATA wsa_data;
     int MySocket;
